tidied up 'update_gx_data_docs.py'
